initial commit
You are working on a data science project, and five samples have been taken from a river at different locations. Each of these water samples has been tested for a particular toxin in the river. After the tests, results have been produced that show the levels of these toxins as a set of five decimal numbers, indicating the toxin levels in each area. As part of this assignment, you are expected to take these five decimal numbers and perform various calculations. The first calculation is to find the mean/average of these five numbers.